PlatRec_Service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.186
opencv-python==4.8.1.78
easyocr==1.7.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0